requirements.txt
einops
accelerate==0.30.1
deepspeed==0.9.5
pillow
peft==0.12.0
transformers==4.44.0
pickle
bitsandbytes==0.44.0
tokenizers==0.19.0
wandb
pandas
scipy
tensorboard